第2轮评测 - 1、/api/dialogue/list、/history、/current_branch 希望都。

+++ schemas.py
@@ -144,7 +144,7 @@ class SkillResponse(BaseModel):
     id: int
     name: str
     description: str
-    damage: int
+    damage: float
     mp_cost: int
     cooldown: int
     current_cooldown: int
